% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_outliers}
\alias{get_outliers}
\title{get_outliers}
\usage{
get_outliers(data)
}
\arguments{
\item{data}{input vector}
}
\value{
outliers in a vector
}
\description{
function to get outliers in a vector
}
\examples{

}
