% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{largest}
\alias{largest}
\title{largest}
\usage{
largest(input)
}
\arguments{
\item{input}{input vector}
}
\value{
maximum number
}
\description{
function to get maximum element in R vector.
}
\examples{
v=c(10,50,5,3)
smallest(v)

50
}
