% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalize_decimal}
\alias{normalize_decimal}
\title{normalize_decimal}
\usage{
normalize_decimal(data)
}
\arguments{
\item{data}{input vector}
}
\value{
normalized vector
}
\description{
function to normalize vector according to largest number digits newNumber=oldNumber/10^digits where digits is largest number digits
}
