% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalize_min_max}
\alias{normalize_min_max}
\title{normalize_min_max}
\usage{
normalize_min_max(data, newMin_newMax)
}
\arguments{
\item{data}{input vector}

\item{newMin_newMax}{vector contains 2 values only newMin and newMax which is your new range to normalize data}
}
\value{
normalized vector
}
\description{
function to normalize data between some range(newMin,newMax)
}
