% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalize_z_score}
\alias{normalize_z_score}
\title{normalize_z_score}
\usage{
normalize_z_score(data)
}
\arguments{
\item{data}{input vector}
}
\value{
normalized vector
}
\description{
function to normalize vector according to z-score
}
