% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_outliers}
\alias{remove_outliers}
\title{remove_outliers}
\usage{
remove_outliers(data)
}
\arguments{
\item{data}{input vector}
}
\value{
vector without outliers
}
\description{
function to remove outliers from a vector
}
\examples{

}
