% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smallest}
\alias{smallest}
\title{smallest}
\usage{
smallest(input)
}
\arguments{
\item{input}{input vector}
}
\value{
minumum number
}
\description{
function to get minumum element in R vector.
}
\examples{
v=c(10,50,5,3)
smallest(v)

3
}
